requirements.txt
# requirements.txt

pinecone
sentence-transformers
python-dotenv
openai
streamlit
elevenlabs
streamlit-audiorec
streamlit-agraph